feat: restructuring in anticipation of old and new flows next to each other with feature flag. With scripts check if the queue names stay the same
